Remove explict solver steps to test purely QS solver convergence
Eliminate all boundary conditions except the ones related to implementing
far-field stresses. For the displacement fields in the outer boundary
used to model 10Mpa of isotropic horizontal stresses, reduce the magnitude
to 1/10 and increase the load steps from 1 to 10. This is to test if the
normal QuasiStatic solver scheme will be sufficient for converging in this
scenario.