Several changes to implement time-dependent QS BC and untrack unnecessary changes
Add an asymptotic, time-dependent scaling factor to insitu stress displacement
boundary condition fields, in order to avoid step input type boundary condition.
This type of application should help with convergence better:

{QS_FINAL_TIME} is determine by how close to 1 the scaling factor needs to be
with the exponential time factor used. Trial and error with this time factor
may be needed. Number of load steps --> the more the better for convergence, but
the trade off is that the longer computational time. This also may need to be
tweaked via trial and error.

Mesh size is changed to 45.0001 cm and other quantities such as horizon, critical
stretch and boundary conditions are corrected accordingly. This mesh size may keep
changing for trial and error with QS solver, but the other files that make note of
changes will also stay in sync.

Delete tracking of some useless files so that there's less clutter in commits and
branches.